Move team and group styles into shared styles. Make buttons lighter and switch to darker color when hovered over.
