+++ src/team/components/UserList.js
@@ -3,7 +3,7 @@ import Radium from "radium";
 
 import Grid from "react-bootstrap/lib/Grid";
 import UserLabel from "./UserLabel";
-import styles from "~/team/styles";
+import styles from "~/shared/styles/userList";
 import { connect } from "react-redux";
 
 @Radium
