--- src/team/styles.js
@@ -1,25 +1,0 @@
-export default {
-    wideBody: {
-        top: "40px",
-        height: "calc(100% - 40px)",
-        textAlign: "center",
-    },
-    teamInfo: {
-        span: {
-            display: "inline-block",
-            verticalAlign: "textTop",
-            marginTop: "40px",
-            marginBottom: "0px",
-            marginLeft: "20px",
-            marginRight: "20px",
-            paddingBottom: "50px",
-        },
-        h1: {
-            paddingTop: "50px",
-            marginTop: "0px",
-        },
-    },
-    memberList: {
-        marginTop: "20px",
-    },
-}
